moved the query opeation to SalesforceBaseSourceConfig. The MultiObject classes extend this class so to access the methods I had to move them from SalesforceSourceConfig.
