+++ src/main/java/io/cdap/plugin/salesforce/plugin/source/batch/SalesforceSourceConfig.java
@@ -17,7 +17,6 @@ package io.cdap.plugin.salesforce.plugin.source.batch;
 
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Strings;
-import com.sforce.async.OperationEnum;
 import com.sforce.soap.partner.PartnerConnection;
 import com.sforce.ws.ConnectionException;
 import io.cdap.cdap.api.annotation.Description;
@@ -89,11 +88,6 @@ public class SalesforceSourceConfig extends SalesforceBaseSourceConfig {
   @Description("Parent of the Salesforce Object. This is used to enable chunking for history tables or shared objects.")
   private String parent;
 
-  @Name(SalesforceSourceConstants.PROPERTY_OPERATION)
-  @Description("If set to query, the query result will only return current rows. If set to queryAll, all records, includ" +
-          "ing deletes will be sourced")
-  private String operation;
-
   @VisibleForTesting
   SalesforceSourceConfig(String referenceName,
                          @Nullable String consumerKey,
@@ -110,19 +104,18 @@ public class SalesforceSourceConfig extends SalesforceBaseSourceConfig {
                          @Nullable String schema,
                          @Nullable String securityToken,
                          @Nullable OAuthInfo oAuthInfo,
+                         @Nullable String operation,
                          @Nullable Boolean enablePKChunk,
                          @Nullable Integer chunkSize,
-                         @Nullable String parent,
-                         String operation) {
+                         @Nullable String parent) {
     super(referenceName, consumerKey, consumerSecret, username, password, loginUrl,
-          datetimeAfter, datetimeBefore, duration, offset, securityToken, oAuthInfo);
+          datetimeAfter, datetimeBefore, duration, offset, operation, securityToken, oAuthInfo);
     this.query = query;
     this.sObjectName = sObjectName;
     this.schema = schema;
     this.enablePKChunk = enablePKChunk;
     this.chunkSize = chunkSize;
     this.parent = parent;
-    this.operation = operation;
   }
 
   /**
@@ -319,20 +312,6 @@ public class SalesforceSourceConfig extends SalesforceBaseSourceConfig {
   public boolean getEnablePKChunk() {
     return enablePKChunk == null ? false : enablePKChunk;
   }
-
-  public String getOperation() {
-    return operation;
-  }
-
-  public OperationEnum getOperationEnum() {
-    try {
-      return OperationEnum.valueOf(operation.toLowerCase());
-    } catch (IllegalArgumentException ex) {
-      throw new InvalidConfigException("Unsupported value for operation: " + operation,
-              SalesforceSourceConstants.PROPERTY_OPERATION);
-    }
-  }
-
 
   public int getChunkSize() {
     return chunkSize == null ? SalesforceSourceConstants.DEFAULT_PK_CHUNK_SIZE : chunkSize;
