+++ src/test/java/io/cdap/plugin/salesforce/plugin/source/batch/SalesforceSourceConfigBuilder.java
@@ -37,7 +37,7 @@ public class SalesforceSourceConfigBuilder {
   private Boolean enablePKChunk;
   private Integer chunkSize;
   private String parent;
-  private String queryType;
+  private String operation;
 
   public SalesforceSourceConfigBuilder setReferenceName(String referenceName) {
     this.referenceName = referenceName;
@@ -124,15 +124,14 @@ public class SalesforceSourceConfigBuilder {
     return this;
   }
 
-  public SalesforceSourceConfigBuilder setQueryType(String queryType) {
-    this.queryType = queryType;
+  public SalesforceSourceConfigBuilder setOperation(String operation) {
+    this.operation = operation;
     return this;
   }
-
 
   public SalesforceSourceConfig build() {
     return new SalesforceSourceConfig(referenceName, consumerKey, consumerSecret, username, password, loginUrl,
                                       query, sObjectName, datetimeAfter, datetimeBefore, duration, offset, schema,
-                                      securityToken, null, enablePKChunk, chunkSize, parent, queryType);
+                                      securityToken, null, operation, enablePKChunk, chunkSize, parent);
   }
 }
