+++ widgets/SalesforceMultiObjects-batchsource.json
@@ -142,6 +142,25 @@
       "label": "Advanced",
       "properties": [
         {
+          "widget-type": "radio-group",
+          "label": "SOQL Query Operation",
+          "name": "SOQL Query Operation",
+          "widget-attributes": {
+            "layout": "inline",
+            "default": "query",
+            "options": [
+              {
+                "id": "query",
+                "label": "query"
+              },
+              {
+                "id": "queryAll",
+                "label": "queryAll"
+              }
+            ]
+          }
+        },
+        {
           "widget-type": "textbox",
           "label": "SObject Name Field",
           "name": "sObjectNameField",
